[refactoring] step3: 파일 위치 수정 및 패키지 분리
