+++ src/main/kotlin/blackjack_dealer/BlackJackResultBoard.kt
@@ -1,12 +1,11 @@
 package blackjack_dealer
 
 import blackjack_dealer.domain.Dealer
-import blackjack_dealer.entity.DealerResult
-import blackjack_dealer.entity.ParticipantResult
-import blackjack_dealer.entity.ParticipantResultState
 import blackjack_dealer.entity.Participants
-import blackjack_dealer.entity.TotalResult
-import blackjack_dealer.ui.OutputView
+import blackjack_dealer.entity.result.DealerResult
+import blackjack_dealer.entity.result.ParticipantResult
+import blackjack_dealer.entity.result.TotalResult
+import blackjack_dealer.entity.state.ParticipantResultState
 
 object BlackJackResultBoard {
     fun getBlackJackResult(dealer: Dealer, participants: Participants): TotalResult {
