+++ src/main/kotlin/blackjack_dealer/domain/Participant.kt
@@ -3,7 +3,7 @@ package blackjack_dealer.domain
 import blackjack_dealer.CardGenerator
 import blackjack_dealer.entity.CardDeque
 import blackjack_dealer.entity.GamerCards
-import blackjack_dealer.entity.GamerCurrentState
+import blackjack_dealer.entity.state.GamerCurrentState
 import blackjack_dealer.entity.toGamerCards
 
 data class Participant(
