--- src/main/kotlin/blackjack_dealer/entity/ParticipantResult.kt
@@ -1,6 +1,0 @@
-package blackjack_dealer.entity
-
-data class ParticipantResult(
-    val name: String,
-    val resultState: ParticipantResultState
-)
