+++ src/main/kotlin/blackjack_dealer/entity/card/CardNumber.kt
@@ -1,4 +1,4 @@
-package blackjack_dealer.entity
+package blackjack_dealer.entity.card
 
 enum class CardNumber(val number: Int) {
     A(1), TWO(2), THREE(3), FOUR(4), FIVE(5), SIX(6), SEVEN(7), EIGHT(8), NINE(9), TEN(10),
