+++ src/main/kotlin/blackjack_dealer/entity/card/CardShape.kt
@@ -1,4 +1,4 @@
-package blackjack_dealer.entity
+package blackjack_dealer.entity.card
 
 enum class CardShape(val shapeName: String) {
     SPADE("스페이드"), HEART("하트"), CLOVER("클로버"), DIAMOND("다이아");
