+++ src/main/kotlin/blackjack_dealer/entity/result/ParticipantResult.kt
@@ -1,0 +1,8 @@
+package blackjack_dealer.entity.result
+
+import blackjack_dealer.entity.state.ParticipantResultState
+
+data class ParticipantResult(
+    val name: String,
+    val resultState: ParticipantResultState
+)
