+++ src/main/kotlin/blackjack_dealer/entity/result/TotalResult.kt
@@ -1,4 +1,4 @@
-package blackjack_dealer.entity
+package blackjack_dealer.entity.result
 
 data class TotalResult(
     val dealerResult: DealerResult,
